Create simple FOR loop
This is a short piece of Python code to illustrate the operation of a FOR loop in Python.  It demonstrates the assignment of variables, the calculation of a set of values from a FOR loop structure, illustrates the concept of iteration of a calculation for which the variable has incrementally changed values, and finally, shows how to use a PRINT string command to display the results.